Create GetWorkingDate
So this is an attempt to check yesterday date to see if it is a non-working day, a Saturday or a Sunday, because the interprise does not work on Week-ends. If it fills up those contitions , it returns none, else, it returns Yesterday in the format DDMMYY. My thiny knowledge of Python had me get inspired with Nigeria and Burundi to create the Cameroon class. But I cannot get it to work properly. Any comment advice and code proposals are welcome...